Move serialized objects to separate file
