+++ python_modules/dagster/dagster/_core/definitions/auto_materialize_rule_evaluation.py
@@ -2,7 +2,6 @@ from abc import ABC, abstractproperty
 from collections import defaultdict
 from enum import Enum
 from typing import (
-    TYPE_CHECKING,
     AbstractSet,
     Dict,
     FrozenSet,
@@ -31,15 +30,13 @@ from dagster._serdes.serdes import (
 )
 from dagster._utils.security import non_secure_md5_hash_str
 
+from .declarative_scheduling.serialized_objects import (
+    AssetConditionEvaluation,
+    AssetConditionEvaluationWithRunIds,
+    AssetConditionSnapshot,
+    AssetSubsetWithMetadata,
+)
 from .partition import PartitionsDefinition, SerializedPartitionsSubset
-
-if TYPE_CHECKING:
-    from .declarative_scheduling.legacy.asset_condition import (
-        AssetConditionEvaluation,
-        AssetConditionEvaluationWithRunIds,
-        AssetConditionSnapshot,
-        AssetSubsetWithMetadata,
-    )
 
 
 @whitelist_for_serdes
@@ -148,7 +145,7 @@ def deserialize_auto_materialize_asset_evaluation_to_asset_condition_evaluation_
     """Provides a backcompat layer to allow deserializing old AutoMaterializeAssetEvaluation
     objects into the new AssetConditionEvaluationWithRunIds objects.
     """
-    from .declarative_scheduling.legacy.asset_condition import (
+    from .declarative_scheduling.serialized_objects import (
         AssetConditionEvaluationWithRunIds,
     )
 
@@ -215,8 +212,8 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
     def _asset_condition_snapshot_from_rule_snapshot(
         self, rule_snapshot: AutoMaterializeRuleSnapshot
     ) -> "AssetConditionSnapshot":
-        from .declarative_scheduling.legacy.asset_condition import AssetConditionSnapshot
         from .declarative_scheduling.legacy.rule_condition import RuleCondition
+        from .declarative_scheduling.serialized_objects import AssetConditionSnapshot
 
         unique_id_parts = [rule_snapshot.class_name, rule_snapshot.description]
         unique_id = non_secure_md5_hash_str("".join(unique_id_parts).encode())
@@ -236,9 +233,7 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         rule_snapshot: AutoMaterializeRuleSnapshot,
     ) -> "AssetConditionEvaluation":
-        from .declarative_scheduling.legacy.asset_condition import (
-            AssetConditionEvaluation,
-            AssetSubsetWithMetadata,
+        from .declarative_scheduling.serialized_objects import (
             HistoricalAllPartitionsSubsetSentinel,
         )
 
@@ -281,14 +276,12 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         is_partitioned: bool,
         decision_type: AutoMaterializeDecisionType,
     ) -> Optional["AssetConditionEvaluation"]:
-        from .declarative_scheduling.legacy.asset_condition import (
-            AssetConditionEvaluation,
-            AssetConditionSnapshot,
-            HistoricalAllPartitionsSubsetSentinel,
-        )
         from .declarative_scheduling.operators.boolean_operators import (
             NotAssetCondition,
             OrAssetCondition,
+        )
+        from .declarative_scheduling.serialized_objects import (
+            HistoricalAllPartitionsSubsetSentinel,
         )
 
         partition_subsets_by_condition_by_rule_snapshot = defaultdict(list)
@@ -376,12 +369,10 @@ class BackcompatAutoMaterializeAssetEvaluationSerializer(PydanticModelSerializer
         whitelist_map: WhitelistMap,
         context: UnpackContext,
     ) -> "AssetConditionEvaluationWithRunIds":
-        from .declarative_scheduling.legacy.asset_condition import (
-            AssetConditionEvaluation,
-            AssetConditionSnapshot,
+        from .declarative_scheduling.operators.boolean_operators import AndAssetCondition
+        from .declarative_scheduling.serialized_objects import (
             HistoricalAllPartitionsSubsetSentinel,
         )
-        from .declarative_scheduling.operators.boolean_operators import AndAssetCondition
 
         asset_key = cast(AssetKey, unpacked_dict.get("asset_key"))
         partition_subsets_by_condition = cast(
