+++ python_modules/dagster/dagster/_core/storage/legacy_storage.py
@@ -13,7 +13,7 @@ from dagster import (
     _check as check,
 )
 from dagster._config.config_schema import UserConfigSchema
-from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import (
+from dagster._core.definitions.declarative_scheduling.serialized_objects import (
     AssetConditionEvaluationWithRunIds,
 )
 from dagster._core.definitions.events import AssetKey
