+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/asset_condition_tests/test_asset_condition.py
@@ -1,6 +1,8 @@
 from dagster import AutoMaterializePolicy, Definitions, asset
 from dagster._core.definitions.declarative_scheduling.legacy.asset_condition import (
     AssetCondition,
+)
+from dagster._core.definitions.declarative_scheduling.serialized_objects import (
     AssetConditionEvaluation,
 )
 from dagster._core.remote_representation.external_data import external_repository_data_from_def
